Use "present" metric to obtain player counts
The present metric is actually formatted like this, with a separate line
per player:
present{name="playername",} 1.0
Thus, when retrieving this metric from the get() function, the value 1.0
will always be returned if any players are online, and 0.0 if there are
none. As stop() only checks if the number of players is greater than 0,
this will still work as intended.